changed it to be one integer

+++ challange_4/main.py
@@ -23,29 +23,26 @@ def main(input_file, output_file):
     nums_file = open(input_file, 'r')
     out = open(output_file, 'w')
     
-    nums = []
+    num = 0
 
     lines = nums_file.readlines()
     for line in lines:
         x = 0
         number =''
-        
+     
         while line[x:x+1].isdigit():
             number += line[x]
             x+=1
-        nums.append(int(number))
-    
-    nums_file.close()
-    
+	if(number != ''):
+	    num = int(number)
+	    nums_file.close()
+	    break 
     out.truncate()
 
-    num_bits = []
-    for number in nums:
-        num_bits.append(num_of_bits(number))
+    num_bits = num_of_bits(num)
     
-    for n in num_bits:
-        out.write(str(n))
-        out.write('\r\n')
+    out.write(str(num_bits))
+    out.write('\r\n')
     
 if __name__ == "__main__":
     input_file = ''
